trosku lepsie pocitanie 2-pakovania

+++ impl/experiment/pp2p.hpp
@@ -8,7 +8,7 @@ using namespace std;
 #define For(i, n) for (int i = 0; i < (int) n; i++)
 
 typedef long long ll;
-typedef vector<int> vi;
+typedef vector<ll> vi;
 typedef vector<vi> vvi;
 typedef pair<int, int> pii;
